ROFO-131 refactor: SearchTopBar를 공통 UI로 분리
