+++ presentation/src/main/java/com/weit2nd/presentation/ui/search/SearchScreen.kt
@@ -3,8 +3,6 @@ package com.weit2nd.presentation.ui.search
 import android.widget.Toast
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.interaction.MutableInteractionSource
-import androidx.compose.foundation.interaction.collectIsFocusedAsState
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
@@ -16,42 +14,36 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.itemsIndexed
-import androidx.compose.foundation.text.BasicTextField
-import androidx.compose.foundation.text.KeyboardActions
-import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
-import androidx.compose.material3.TextFieldDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.input.ImeAction
-import androidx.compose.ui.text.input.VisualTransformation
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.weit2nd.domain.model.search.Place
 import com.weit2nd.presentation.R
+import com.weit2nd.presentation.navigation.dto.PlaceSearchDTO
+import com.weit2nd.presentation.ui.common.SearchTopBar
 import org.orbitmvi.orbit.compose.collectAsState
 import org.orbitmvi.orbit.compose.collectSideEffect
 
 @Composable
 fun SearchScreen(
     vm: SearchViewModel = hiltViewModel(),
-    navToMap: () -> Unit,
+    navToHome: (PlaceSearchDTO) -> Unit,
     navToBack: () -> Unit,
 ) {
     val state by vm.collectAsState()
@@ -63,6 +55,9 @@ fun SearchScreen(
             }
             is SearchSideEffect.ShowToastMessage -> {
                 Toast.makeText(context, sideEffect.message, Toast.LENGTH_SHORT).show()
+            }
+            is SearchSideEffect.NavToHome -> {
+                navToHome(sideEffect.placeSearch)
             }
         }
     }
@@ -90,39 +85,6 @@ fun SearchScreen(
             onHistoryClick = vm::onHistoryClick,
             onHistoryRemove = vm::onHistoryRemove,
             onPlaceClick = vm::onSearchResultClick,
-        )
-    }
-}
-
-@Composable
-fun SearchTopBar(
-    modifier: Modifier = Modifier,
-    textFieldEnabled: Boolean = true,
-    searchWords: String,
-    onClear: () -> Unit,
-    onSearchButtonClick: () -> Unit,
-    onSearchWordsChanged: (String) -> Unit,
-    onNavigationButtonClick: () -> Unit,
-) {
-    Row(
-        modifier = modifier,
-        verticalAlignment = Alignment.CenterVertically,
-    ) {
-        IconButton(
-            onClick = onNavigationButtonClick,
-        ) {
-            Icon(
-                painter = painterResource(id = R.drawable.ic_navigate),
-                contentDescription = "",
-            )
-        }
-        SearchTextField(
-            modifier = Modifier.weight(1f),
-            enabled = textFieldEnabled,
-            searchWords = searchWords,
-            onSearchWordsChanged = onSearchWordsChanged,
-            onClear = onClear,
-            onSearchButtonClick = onSearchButtonClick,
         )
     }
 }
@@ -158,9 +120,9 @@ private fun SearchContent(
             }
             SearchHistoryItem(
                 modifier =
-                Modifier
-                    .fillMaxSize()
-                    .padding(horizontal = 8.dp),
+                    Modifier
+                        .fillMaxSize()
+                        .padding(horizontal = 8.dp),
                 history = history,
                 onClick = {
                     onHistoryClick(history)
@@ -199,9 +161,9 @@ private fun SearchContent(
             }
             SearchPlaceItem(
                 modifier =
-                Modifier
-                    .fillMaxWidth()
-                    .padding(8.dp),
+                    Modifier
+                        .fillMaxWidth()
+                        .padding(8.dp),
                 name = place.placeName,
                 address = place.roadAddressName.takeIf { it.isNotEmpty() } ?: place.addressName,
                 onClick = {
@@ -209,84 +171,6 @@ private fun SearchContent(
                 },
             )
         }
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-private fun SearchTextField(
-    modifier: Modifier = Modifier,
-    enabled: Boolean = true,
-    searchWords: String,
-    onSearchWordsChanged: (String) -> Unit,
-    onClear: () -> Unit,
-    onSearchButtonClick: () -> Unit,
-) {
-    val interactionSource =
-        remember {
-            MutableInteractionSource()
-        }
-    val hasFocus by interactionSource.collectIsFocusedAsState()
-
-    BasicTextField(
-        modifier = modifier,
-        value = searchWords,
-        onValueChange = onSearchWordsChanged,
-        interactionSource = interactionSource,
-        keyboardOptions =
-            KeyboardOptions.Default.copy(
-                imeAction = ImeAction.Search,
-            ),
-        keyboardActions =
-            KeyboardActions(
-                onSearch = {
-                    onSearchButtonClick()
-                },
-            ),
-    ) { innerTextField ->
-        TextFieldDefaults.DecorationBox(
-            value = searchWords,
-            innerTextField = innerTextField,
-            enabled = enabled,
-            singleLine = true,
-            visualTransformation = VisualTransformation.None,
-            interactionSource = remember { MutableInteractionSource() },
-            placeholder = {
-                Text(
-                    text = "장소를 입력해!",
-                )
-            },
-            leadingIcon = {
-                Icon(
-                    painter = painterResource(id = R.drawable.ic_search_glass),
-                    contentDescription = "장소 검색 바",
-                )
-            },
-            trailingIcon = {
-                if (searchWords.isNotEmpty() && hasFocus) {
-                    IconButton(
-                        onClick = onClear,
-                    ) {
-                        Icon(
-                            painter = painterResource(id = R.drawable.ic_input_delete_filled),
-                            contentDescription = "검색어 정리",
-                            tint = Color.Unspecified,
-                        )
-                    }
-                }
-            },
-            // TODO Material Theme를 적용하고 나면 제거
-            colors =
-                TextFieldDefaults.colors(
-                    disabledContainerColor = Color.White,
-                    errorContainerColor = Color.White,
-                    focusedContainerColor = Color.White,
-                    unfocusedContainerColor = Color.White,
-                    unfocusedIndicatorColor = Color.White,
-                    disabledIndicatorColor = Color.White,
-                    focusedIndicatorColor = Color(0xFF555555),
-                ),
-        )
     }
 }
 
@@ -339,11 +223,11 @@ private fun SearchHistoryItem(
     ) {
         Text(
             modifier =
-            Modifier
-                .weight(1f)
-                .clickable {
-                    onClick()
-                },
+                Modifier
+                    .weight(1f)
+                    .clickable {
+                        onClick()
+                    },
             text = history,
             fontSize = 18.sp,
         )
@@ -410,38 +294,11 @@ private fun SearchContentPreview() {
 private fun SearchHistoryItemPreview() {
     SearchHistoryItem(
         modifier =
-        Modifier
-            .fillMaxWidth()
-            .background(Color.White),
+            Modifier
+                .fillMaxWidth()
+                .background(Color.White),
         history = "명륜진사",
         onClick = {},
         onRemove = {},
-    )
-}
-
-@Preview
-@Composable
-private fun SearchTextFieldPreview() {
-    SearchTextField(
-        modifier = Modifier.fillMaxWidth(),
-        searchWords = "안녕하세요",
-        onSearchWordsChanged = {},
-        onClear = {},
-        onSearchButtonClick = {},
-    )
-}
-
-@Preview
-@Composable
-private fun SearchTopBarPerview() {
-    SearchTopBar(
-        modifier = Modifier
-            .fillMaxWidth()
-            .background(Color.White),
-        searchWords = "감사해요",
-        onClear = {},
-        onSearchButtonClick = {},
-        onSearchWordsChanged = {},
-        onNavigationButtonClick = {},
     )
 }
